Add support for downloading and viewing files
Adds options "download" and "view" to command. "download" is a path,
absolute or relative to SSH login directory (${HOME}), of a file which
will be downloaded via SFTP to the Android download directory after
executing command.

You may set command to a no-op (e.g. "cmd" : "echo") to simply
download a file. "cmd" : "" is illegal (since "cmd" is mandatory) and
"cmd" : ";" fails since "cmd" will be passed as argument to nohup.

Otherwise combine "download" with "wait" : true in order to download the
result of a command execution.

The progress of the download will be displayed as notification. With
"view" : true, the notification will be cancelled on success of the
download and an ACTION_VIEW intent will be invoked immediately with the
mime type derived from the file extension. With "view" : false, the user
may start the ACTION_VIEW intent by tapping on the notification.

The documentation is updated accordingly. The description of the option
"wait" is fixed.

Remark: Currently the file name specified in download is fixed. It might
be a nice extension to make it depend on the command output or the
shell environment after command execution. It's not necessary for my
personal needs so far, so feel free to implement it yourself.