Add support for rulers
A ruler represents a settable integer property. Its value is in the range
from min to max and may be set in intervals of step. getcmd is used for
retrieving the value of the property from server and setcmd is used for
changing the property on the server. getcmd must output the value to the
first line of stdout, setcmd must set the property to the content of the
nvironment variable VALUE.

The ruler is displayed as a SeekBar. When displaying the server page,
the getcmd is invoked in order to get the value of the ruler. The value
may be updated by clicking (touching) the command list entry (outside
the SeekBar). When the SeekBar is changed, the setcmd is called
immediately in order to set the property on the server.

The ruler entry accepts most of the attributes of command as well. See
help or the source code for more details.

Polymorphism is used in order to distinguish between rulers and commands
in the "commands" attribute of the server JSON object.

Also some missing strings are added.